requirements.txt
groq
python-dotenv
fastapi
uvicorn[standard]
langchain
langchain-groq
langchain-community
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
requests
pydantic